requirements.txt
altair==5.5.0
confluent_kafka==2.10.0
kafka_python==2.2.10
pandas==2.2.3
Requests==2.32.3
streamlit==1.44.1
